requirements.txt
torch
transformers
chromadb
sentence-transformers
tqdm
diffusers
accelerate
langdetect
Pillow
gradio
